small bug fixes

+++ script/evaluation/generate_results.py
@@ -47,7 +47,8 @@ def generate_results(
     results_dir = f"../evaluation/{run_name}/predictions/"
     os.makedirs(results_dir, exist_ok=True)
     filename = os.path.join(results_dir, "results.csv")
-
+    if os.path.exists(filename):
+        raise RuntimeError("output file already exists")
     write_header = not os.path.exists(filename)
     if write_header:
         pd.DataFrame(columns=["labels", "output", "path", "tile", "patient"]).to_csv(
